trip accept fix

+++ lib/screens/driver/driver_home/repository/driver_repository.dart
@@ -21,7 +21,7 @@ class DriverRepository {
     });
   }
 
-  Future<void> AcceptTrip(String docId, String newDriverId, double rent) async {
+  Future<void> AcceptTrip(String docId, String newDriverId, int rent) async {
     try {
       await _firestore.collection('All Trips').doc(docId).update({
         'driverId': newDriverId,
